Сделал вывод расписание у студентов, указание группы. Детализировал меню.

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->integer('telegram_id');
+            $table->bigInteger('telegram_id');
             $table->string('telegram_username')->nullable();
             $table->string('telegram_first_name');
             $table->string('telegram_last_name')->nullable();
